Moved PUT function into the same file as DELETE
