--- .env
@@ -1,5 +1,0 @@
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=postgres
-POSTGRES_DB=nycmesh
-POSTGRES_HOST=localhost
-POSTGRES_PORT=5432
